% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aggregate.rora.R
\name{aggregate.rora}
\alias{aggregate.rora}
\title{Compute Recharge}
\usage{
\method{aggregate}{rora}(x, by = "months", ...)
}
\arguments{
\item{x}{an object of class "rora."}

\item{by}{the time period to aggregate by. See \bold{Details}.}

\item{\dots}{not used, required for other methods.}
}
\value{
The recharge for each period specified in \code{by}. The units are
the same as for \code{x}.
}
\description{
Computes recharge from an object of class "rora" from a streamflow recession
displacement analysis.
}
\details{
The argument \code{by} can be either a character indicating the period, or a list
created by \code{setSeasons}. If a character , thne must be "months," "years,"
"calendar years," "water years," or "climate years." May be abbreviated; and
"years" is the same as "calendar years."
}
\examples{
\dontrun{
library(smwrData)
data(GlacialRidge)
G12.rise <- with(GlacialRidge, rise(G12, datetime, MPelev=1126.42, STAID="G12"))
# monthly summary of recharge in feet
aggregate(G12.rise)
}
}
\seealso{
\code{\link{rora}}, \code{\link{setSeasons}}
}
\keyword{recharge}

