% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.fall.R
\name{confirm.fall}
\alias{confirm.fall}
\title{Confirm Groundwater Recessions}
\usage{
\method{confirm}{fall}(x, all = FALSE, before = 3, after = 3, ...)
}
\arguments{
\item{x}{an object of class "fall."}

\item{all}{accept all recessions without review?}

\item{before}{begin review plot \code{before} days prior to recession.}

\item{after}{end review plot \code{after} days after the end of the recession.}

\item{\dots}{not used, required for other methods.}
}
\value{
The object \code{x} is returned with updated recessions
}
\description{
This function processes the output of function \code{fall} and requires the user to 
accept or reject the recession.
}
\keyword{manip}

