% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm.R
\name{confirm}
\alias{confirm}
\title{Confirm an Analysis}
\usage{
confirm(x, ...)
}
\arguments{
\item{x}{the object to be confirmed}

\item{\dots}{additional arguments required for specific methods}
}
\value{
The object returned depends on the specific method.
}
\description{
Review and accept the results of an analysis
}
\keyword{manip}
