% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistentFDC.R
\name{consistentFDC}
\alias{consistentFDC}
\title{Flow-Duration Curve}
\usage{
consistentFDC(x)
}
\arguments{
\item{x}{a vector of the 13 estimated points on the flow-duration curve from regresion regressions
in Minnesota.}
}
\value{
a vector of 13 consitently decreasing estimates.
}
\description{
Construct a consistent flow-duration curve from the 13 estiamtes from a regional regression.
}

