% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dvStat.R
\name{dvStat}
\alias{dvStat}
\title{Daily Value Statistics}
\usage{
dvStat(x, Dates, Start = NULL, End = NULL, by = NULL, pre = NULL,
  stat = min, na.rm = TRUE, STAID = "Unknown", ...)
}
\arguments{
\item{x}{the daily value data to be summarized. Missing values are permitted, but
some summary statistics could be biased if missing data are present.}

\item{Dates}{the date for each \code{x}, should be of class "Date." Missing values
are not permitted.}

\item{Start}{the start date for the analysis, can be either a character string or
class "Date."}

\item{End}{the end date for the analysis, can be either a character string or
class "Date."}

\item{by}{the grouping variable, generally the output from some date grouping function
like \code{waterYear}, \code{climateYear}, or \code{seasonYear}. The default is to construct
a group of the full range of data in calendar years.}

\item{pre}{any preprocessing function like \code{movingAve}. The function must return
a vector as long as the input vector.}

\item{stat}{the statistical function. Must accept the \code{na.rm} argument and return
a single value.}

\item{na.rm}{remove missiing values option for the function \code{stat}.}

\item{STAID}{the station identifier for the data.}

\item{\dots}{any additional arguments for the function \code{pre}.}
}
\value{
A data.frame containing the \code{STAID}, groups generated by
\code{by}, the number of non-missing values in the grouop, and the statistic 
and first date on which that statistic occurred.
}
\description{
Compute summary statistics for daily values, possibly grouped by
years or seasons.
}
\examples{

\dontrun{
# For these examples, print to console
.pager <- options("pager")
options(pager="console")
# See the demo for examples of how to use the dvStat function.
demo(package="DVstats")
options(.pager)
}
}
