\docType{package}
\name{DVstats-package}
\alias{DVstats}
\alias{DVstats-package}
\title{Functions for manipulating daily values}
\description{
  This package has specialized functions for managing or
  manipulating daily hydrologic data.
}
\details{
  \tabular{ll}{ Package: \tab DVstats\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 2013-06-13\cr
  License: \tab File LICENSE\cr Depends: \tab lubridate,
  USGSwsBase, USGSwsGraphs, USGSwsStats, robust\cr } This
  package contains functions that manage or manipipulate
  daily hydrologic data.
}
\author{
  Dave Lorenz <lorenz@usgs.gov>

  Maintainer: Dave Lorenz <lorenz@usgs.gov>
}
\references{
  Lorenz, D.L., in preparation, DVstats---an R package for
  managing daily hydrologic data, version 0.1
}
\seealso{
  \code{\link[USGSwsData:USGSwsData-package]{USGSwsData}}
}
\keyword{package}

