% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estFDC.R
\name{estFDC}
\alias{estFDC}
\title{Estimate FDC points}
\usage{
estFDC(site.data, region, expand = TRUE)
}
\arguments{
\item{site.data}{a data frame consisting of one row and the correct columns
of basin characteristics for the region}

\item{region}{a character string indicating the region that contains the ungaged site,
must be one of "A," "BC," "D," "E," or "F."}

\item{expand}{logical value, if \code{TRUE}, then expand the computed points so that
a graph can be prepared, otherwise only the 13 estimated points.}
}
\value{
A data frame containing the columns Probs, the exceedance probabilities,
and FDC, the estiamte streamflow at the specified probability level.
}
\description{
Estimate the streamflow at selected probability of exceedance values for
ungaged sites in Minnesota.
}
\note{
The selected points on the FDC can also be obtained using StreamStats for
Minnesota.
}

