% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend.R
\name{extend}
\alias{extend}
\title{Series Extention}
\usage{
extend(x)
}
\arguments{
\item{x}{a numeric vector that represent a regular series. Missing values are permitted and
are replaced with non-missing values.}
}
\value{
A vector like \code{x} with missing value replaced.
}
\description{
Extend a regular series with missing values at the beginning or end by replacing
the beginning missing values with the first non-missing value and replacing
the ending missing values with the last non-missing value
}
\note{
Interior missing values should be replaced first with a function like
\code{fillMissing}.
}
\examples{
XT <- c(NA, NA, 1.2, 2.4, 3.5, NA)
extend(XT)
}
\seealso{
\code{\link{fillMissing}}
}
\keyword{manip}

