% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fall.R
\name{fall}
\alias{fall}
\title{Recessions}
\usage{
fall(x, Dates, Start = NULL, End = NULL, by = NULL, MPelev = NULL,
  min.duration = 10, STAID = "Unknown")
}
\arguments{
\item{x}{the daily value data to be summarized. Missing values are not permitted
within the time specified by \code{Start} and \code{End}.}

\item{Dates}{the date for each \code{x}, should be of class "Date." Missing values
are not permitted.}

\item{Start}{the start date for the analysis, can be either a character string or
class "Date."}

\item{End}{the end date for the analysis, can be either a character string or
class "Date."}

\item{by}{the months to search for recessions, can be either month numbers or abbreviations.
The default is use all months.}

\item{MPelev}{the measuring point elevation. Required if \code{x} is a depth below land
surface or other measuring point.}

\item{min.duration}{the minimum length for a recession to be included in the output.}

\item{STAID}{the station identifier for the data.}
}
\value{
an object of class "fall" containing a data frame of the selected data, a data
frame of the recession information, and other information about the analysis.
}
\description{
Identify groundwater recessions
}
\examples{
\dontrun{
library(smwrData)
data(GlacialRidge)
with(GlacialRidge, fall(G12, datetime, MPelev=1126.42))
}
}
\keyword{recession}
