% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqReport.R
\name{freqReport}
\alias{freqReport}
\title{Create Frequency Analysis Report}
\usage{
freqReport(x, file)
}
\arguments{
\item{x}{and object of class "freqAnal."}

\item{file}{the output file base name; the .pdf suffix 
is appended to make the actual file name. if missing, then the
name of \code{x} is used as the base name.}
}
\value{
The actual file name is returned invisibly.
}
\description{
Create a 2-page pdf file report of a frequency analysis. The 
report contains the text output from \code{print(x)}, the estimated
values of selected probabilities of exceedance, and a diagnostic plot
for each fitted distribution.
}
