\name{hmean}
\alias{hmean}
\title{Harmonic Mean}
\usage{
hmean(x, na.rm = FALSE)
}
\arguments{
  \item{x}{the data to compute the harmonic mean.}

  \item{na.rm}{a logical value indicating whether NA values
  should be removed before computation.}
}
\value{
The harmonic mean of \code{x}.
}
\description{
Compute the harmonic mean accoring to the EPA DFLOW manual.
}
\references{
Rossman, L.A., No Date, DFLOW user's manual: U.S.
Environmental Protection Agency, Cincinnati, Ohio, 26 p.
}

