% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.baseflow.R
\name{plot.baseflow}
\alias{plot.baseflow}
\title{Plot an Object}
\usage{
\method{plot}{baseflow}(x, which = "All", set.up = TRUE, ...)
}
\arguments{
\item{x}{an object of class "baseflow."}

\item{which}{either "All" for the entire hydrogaph, "by year" for calendar year
hydrographs, or the calendar year to plot.}

\item{set.up}{set up the graphics page? Set to \code{FALSE} if the graphics page
has been set up with a call to \code{setPage}.}

\item{\dots}{any additional arguments to \code{timePlot}, other than
\code{x} and \code{y}.}
}
\value{
The object \code{x} is returned invisibly
}
\description{
Plot an object of class "baseflow" from any baseflow analysis.
}
\section{Side Effect}{
 The object \code{x} is plotted
}
\seealso{
\code{\link{bfi}}, \code{\link{hysep}}, \code{\link{part}},
}
\keyword{hplot}

