% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fall.R
\name{plot.fall}
\alias{plot.fall}
\title{Plot an Object}
\usage{
\method{plot}{fall}(x, which = "All", set.up = TRUE, before = 0,
  after = before, color.line = "blue", color.ba = "black", ...)
}
\arguments{
\item{x}{an object of class "fall."}

\item{which}{either "All" for all recessions, or any number of element
numbers (not Index number) of the recessions to plot. If \code{which} is element
numbers, then each recession is plotted individually, otherwise all recessions
are plotted on the same graph.}

\item{set.up}{set up the graphics page? Set to \code{FALSE} if the graphics page
has been set up with a call to \code{setPage}.}

\item{before}{begin plot \code{before} days prior to recession.}

\item{after}{end plot \code{after} days after the end of the recession.
See \bold{Details}.}

\item{color.line}{the color of the recession lines. See \bold{Details}.}

\item{color.ba}{the color of the before and after lines, if drawn.}

\item{\dots}{not used, required for other methods.}
}
\value{
The object \code{x} is returned invisibly
}
\description{
Plot an object of class "rise" from a groundwater recession analysis.
}
\details{
The \code{before} and \code{after} arguments prepend and append those
number of days to the plot of the recession, which can be useful for 
assessing the recession.

If \code{color.line} is a vector of colors, then the rates of the recessions are
classified accoring to the number of colors. The smallest rates correspond to the 
first color and the largest rates correspond to the last color.
}
\seealso{
\code{\link{fall}}
}
\keyword{hplot}

