% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.freqAnal.R
\name{plot.freqAnal}
\alias{plot.freqAnal}
\title{Plot an Object}
\usage{
\method{plot}{freqAnal}(x, which = "Final", set.up = TRUE, ...)
}
\arguments{
\item{x}{an object of class "freqAnal."}

\item{which}{either "Final," for the data and the fitted line on a normal
probability scale, or the method. See \bold{Details}.}

\item{set.up}{set up the graphics page? Set \code{set.up} to
\code{FALSE} if the graphics page has been set up with a call to
\code{setPage} or other device.}

\item{\dots}{not used, required for other methods.}
}
\value{
The object \code{x} is returned invisibly
}
\description{
Probability plot of an object of class "freqAnal" frequency analysis.
}
\details{
If \code{which} is "Final," then the "traditional" probability plot
is created that shows the non-zero values adjusted for any
zero values and the fitted line.
Otherwise \code{which} must be one of "default," "emp," "lp3," "ln3,"
or "lgev." In that case, the nonzero data are plotted on the
corresponding probability scale (uniform for "emp"). If \code{which} is
"default," then the selected prediction method is plotted.
}
\section{Side Effect}{
 The object \code{x} is plotted
}
\seealso{
\code{\link{freqAnal}}
}
\keyword{hplot}

