% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.rora.R
\name{plot.rora}
\alias{plot.rora}
\title{Plot an Object}
\usage{
\method{plot}{rora}(x, period = "quarter", col = c(2, 4, 5), ...)
}
\arguments{
\item{x}{an object of class "rora"}

\item{period}{character string indicating the grouping of the graphs.
Must be a valid argument to the \code{by} arguemnt of \code{seq.Date}.}

\item{col}{colors.}

\item{\dots}{not used, required for other methods.}
}
\value{
The object \code{x} is returned invisibly
}
\description{
Plot an object of class "rora" from a streamflow recession
displacement analysis.
}
\seealso{
\code{\link{rora}}
}
\keyword{hplot}

