% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.fall.R
\name{print.fall}
\alias{print.fall}
\title{Print an Object}
\usage{
\method{print}{fall}(x, recessions = FALSE, ...)
}
\arguments{
\item{x}{an object of class "fall."}

\item{recessions}{logical if \code{TRUE}, then print the recession summary, otherwise
the recesisons are not printed}

\item{\dots}{not used, required for other methods.}
}
\value{
The object \code{x} is returned invisibly
}
\description{
Print an object of class "fall" from a groundwater recession analysis.
}
\seealso{
\code{\link{fall}}
}
\keyword{print}

