% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.rora.R
\name{print.rora}
\alias{print.rora}
\title{Print an Object}
\usage{
\method{print}{rora}(x, ...)
}
\arguments{
\item{x}{an object of class "rora"}

\item{\dots}{not used, required for other methods.}
}
\value{
The object \code{x} is returned invisibly
}
\description{
Print an object of class "rora" from a streamflow recession
displacement analysis.
}
\section{Side Effect}{
 The object \code{x} is printed
}
\seealso{
\code{\link{rora}}
}
\keyword{print}

