% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.fall.R
\name{[.fall}
\alias{[.fall}
\title{Subset an Object}
\usage{
\method{[}{fall}(x, i, ...)
}
\arguments{
\item{x}{an object of class "fall."}

\item{i}{index specifying the recession to extract. See \bold{Details}.}

\item{\dots}{not used, required for other methods.}
}
\value{
A data frame of the specified recession containing these columns:\cr
\describe{
\item{Dates}{The date of the observation}
\item{GWLeve}{The ground water level value}
\item{Index}{The recession index number}
\item{Seqno}{The sequence number for each day of the recession}
}
Note that the sequence number of the recession begins at 0 becuase that value
represents the peak and the recession starts the following day.
}
\description{
Extracts or subsets an object of class "fall."
}
\details{
The values for \code{i} can be either a single numeric index or a single 
character string. If \code{i} is numeric, then extract that element number.
If \code{i} is character, then extract that Index number.
}
\examples{
\dontrun{
library(smwrData)
data(GlacialRidge)
# Select the first recession of at least 10 days
with(GlacialRidge, fall(G12, datetime, MPelev=1126.42))[1]
}
}
\seealso{
\code{\link{fall}}
}
\keyword{manip}
