% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wtf.R
\name{wtf}
\alias{wtf}
\title{Water Table Rises}
\usage{
wtf(x, Dates, Start = NULL, End = NULL, MPelev = NULL,
  STAID = "Unknown")
}
\arguments{
\item{x}{the daily value data to be summarized. Missing values are not permitted
within the time specified by \code{Start} and \code{end}.}

\item{Dates}{the date for each \code{x}, should be of class "Date." Missing values
are not permitted.}

\item{Start}{the start date for the analysis, can be either a character string or
class "Date."}

\item{End}{the end date for the analysis, can be either a character string or
class "Date."}

\item{MPelev}{the measuring point elevation. Required if \code{x} is a depth below land
surface or measuring point.}

\item{STAID}{the station identifier for the data.}
}
\value{
an object of class "rise" and inherits class "data.frame" of the selected data, a data
frame of the recession information, and other information about the analysis.
}
\description{
Identify hypothetical groundwater rises based on extrapolated antecedent 
recession rates in hydrographs of daily values.
}
\note{
Healy and Cook (2002) descibe the water-table fluctuation method for estimating
recharge.

The antecedent recession is modeled as log-linear recession, the projected recession is a
fixed fraction of the current recession. The fraction is based on the last 8 days of the recession
or extrapolated if the recession is less than 8 days and at least 4 days in length. If the recession
is less than 4 days in length, the previous recession rate is carried forward. For any rises
occuring before the first recession of at least 4 days, the recession rate is 0, which 
replicates the \code{rise} function.
}
\examples{

\dontrun{
library(smwrData)
data(GlacialRidge)
with(GlacialRidge, wtf(G12, datetime, MPelev=1126.42, STAID="G12"))
}
}
\references{
Healy, R.W., and Cook, P.G., 2002, Using ground-water levels to estimate 
recharge. Hydrogeology Journal, v. 10, p. 91--109.
}
\keyword{recession}

