% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.rora.R
\name{aggregate.rora}
\alias{aggregate.rora}
\title{Compute Recharge}
\usage{
\method{aggregate}{rora}(x, by = "months", ...)
}
\arguments{
\item{x}{an object of class "rora."}

\item{by}{the time period to aggregate by. See \bold{Details}.}

\item{\dots}{not used, required for other methods.}
}
\value{
A data frame reporting the recharge and number of peaks for each period 
specified in \code{by}. The units of recharge are inches.
}
\description{
Computes recharge from an object of class "rora" from a streamflow recession 
displacement analysis.
}
\details{
The argument \code{by} can be either a character indicating the period, or a list
created by \code{setSeasons}. If a character , thne must be "months," "years," 
"calendar years," "water years," or "climate years." May be abbreviated; and 
"years" is the same as "calendar years."
}
\examples{
# See the vignette
}
\seealso{
\code{\link{rora}}, \code{\link{setSeasons}}
}
\keyword{recharge}

