% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/DVstats-package.R
\docType{package}
\name{DVstats-package}
\alias{DVstats}
\alias{DVstats-package}
\title{Functions for manipulating daily values}
\description{
This package has specialized functions for managing or
manipulating daily hydrologic data.
}
\details{
\tabular{ll}{ Package: \tab DVstats\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2015-06-01\cr
License: \tab File LICENSE\cr
Depends: \tab lubridate, smwrBase, smwrGraphs,
smwrStats, robust\cr }
This package contains functions that manage or manipipulate daily
hydrologic data.
}
\author{
Dave Lorenz <lorenz@usgs.gov>

Maintainer: Dave Lorenz <lorenz@usgs.gov>
}
\references{
Lorenz, D.L., in preparation, DVstats---an R package for managing
daily hydrologic data, version X
}
\seealso{
\code{\link[smwrData:smwrData-package]{smwrData}}
}
\keyword{package}

