% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rora.R
\name{plot.rora}
\alias{plot.rora}
\title{Plot an Object}
\usage{
\method{plot}{rora}(x, which = "All", set.up = TRUE, colors = c("none",
  "blue", "cyan"), ...)
}
\arguments{
\item{x}{an object of class "rora"}

\item{which}{either "All" for the entire hydrogaph, "by year" for calendar year
hydrographs, or the calendar year to plot.}

\item{set.up}{set up the graphics page? Set to \code{FALSE} if the graphics page
has been set up with a call to \code{setPage}.}

\item{colors}{a vector of length 3 specifying the colors for the peak and
critical time, the hypothetical streamflow recession, and the calulated
streamflow recession displacement, respectively.}

\item{\dots}{not used, required for other methods.}
}
\value{
The object \code{x} is returned invisibly
}
\description{
Plot an object of class "rora" from a streamflow recession 
displacement analysis.
}
\seealso{
\code{\link{rora}}
}
\keyword{hplot}

