% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PredModels.R
\docType{data}
\name{PredModels}
\alias{PredModels}
\title{Prediction Models
The 13 regional predcition models for the points on a FDC.}
\format{Data frame with  rows and  columns\cr
\tabular{lll}{
Name \tab Type \tab Description\cr
A \tab list \tab The models for region A\cr
BC \tab list \tab The models for region BC\cr
D \tab list \tab The models for region D\cr
E \tab list \tab The models for region E\cr
F \tab list \tab The models for region F\cr
}}
\usage{
PredModels
}
\description{
Prediction Models
The 13 regional predcition models for the points on a FDC.
}
\keyword{datasets}

