% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rise.R
\name{rise}
\alias{rise}
\title{Water Table Rises}
\usage{
rise(x, Dates, Start = NULL, End = NULL, MPelev = NULL,
  STAID = "Unknown")
}
\arguments{
\item{x}{the daily value data to be summarized. Missing values are not permitted
within the time specified by \code{Start} and \code{end}.}

\item{Dates}{the date for each \code{x}, should be of class "Date." Missing values
are not permitted.}

\item{Start}{the start date for the analysis, can be either a character string or
class "Date."}

\item{End}{the end date for the analysis, can be either a character string or
class "Date."}

\item{MPelev}{the measuring point elevation. Required if \code{x} is a depth below land
surface or measuring point.}

\item{STAID}{the station identifier for the data.}
}
\value{
An object of class "rise" and inherits class "data.frame" of the selected data, a data
frame of the recession information, and other information about the analysis.
}
\description{
Identify groundwater rises based on hydrographs of daily values.
}
\note{
Rutledge (2002) describes the rise methed for computing recharge. the method is
a variation on the water-table fluctuation method that assumes no recession.
}
\examples{

\dontrun{
library(smwrData)
data(GlacialRidge)
with(GlacialRidge, rise(G12, datetime, MPelev=1126.42, STAID="G12"))
}
}
\references{
Rutledge, A.T., 2002, User guide for the PULSE program: 
U.S. Geological Survey Open-File Report 02--455, 34 p.
}
\keyword{recession}
